+++ be/src/main/java/com/jjikmuk/sikdorak/auth/service/OAuthApiClient.java
@@ -1,4 +1,4 @@
-package com.jjikmuk.sikdorak.auth.service.kakao;
+package com.jjikmuk.sikdorak.auth.service;
 
 import com.jjikmuk.sikdorak.auth.controller.dto.response.UserInfoResponse;
 import com.jjikmuk.sikdorak.common.config.feignclient.FeignClientHeaderConfiguration;
@@ -6,8 +6,8 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-@FeignClient(name="KakaoApiClient", url = "https://kapi.kakao.com/", configuration = {FeignClientHeaderConfiguration.class})
-public interface KakaoApiClient {
+@FeignClient(name="api-client", url = "https://kapi.kakao.com/", configuration = {FeignClientHeaderConfiguration.class})
+public interface OAuthApiClient {
 
     @GetMapping("/v2/user/me")
     UserInfoResponse getUserInfo(@RequestHeader("Authorization") String accessToken);
