+++ be/src/main/java/com/jjikmuk/sikdorak/auth/service/OAuthTokenClient.java
@@ -1,4 +1,4 @@
-package com.jjikmuk.sikdorak.auth.service.kakao;
+package com.jjikmuk.sikdorak.auth.service;
 
 import com.jjikmuk.sikdorak.auth.controller.dto.response.OAuthTokenResponse;
 import com.jjikmuk.sikdorak.common.config.feignclient.FeignClientHeaderConfiguration;
@@ -6,8 +6,8 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(name="KakaoClient", url = "https://kauth.kakao.com/", configuration = {FeignClientHeaderConfiguration.class})
-public interface KakaoOAuthClient {
+@FeignClient(name="token-client", url = "https://kauth.kakao.com/", configuration = {FeignClientHeaderConfiguration.class})
+public interface OAuthTokenClient {
 
     @PostMapping("/oauth/token")
     OAuthTokenResponse getAccessToken(@RequestParam(value = "grant_type") String grantType,
